minor code revisions in PR
