+++ cmass/conf/nbody/1gpch.yaml
@@ -11,22 +11,18 @@ save_particles: false  # whether to save particle data
 save_velocities: true  # whether to save velocity field data
 
 zi: 20            # initial redshift
-zf: 0.4
-zmax: 0.6         # max redshift of range
-zmin: 0.4         # min redshift of range
-nsnap: 9         # number of snapshots to save
+zf: 0.5           # final redshift
+supersampling: 1  # particles resolution relative to meshgrid
 
-supersampling: 3  # particles resolution relative to meshgrid
 # borglpt, borgpm and pinocchio-only
 transfer: 'CLASS' # transfer function (EH, CLASS, CAMB or SYREN. Only EH or CLASS for borg)
 
 # borglpt-only
 order: 2          # order of the LPT expansion
 
-# pmwd-only
-B: 1.5            # force grid resolution relative to particle grid
-nstep_i: 10        # number of initial PM integration steps
-nstep_snap: 5     # number of PM integration steps between snapshots
+# pmwd and borgpm-only
+B: 2              # force grid resolution relative to particle grid
+N_steps: 20       # number of PM integration steps
 
 # borgpm-only
 COLA: true        # whether to use COLA
