+++ cmass/survey/tools.py
@@ -9,6 +9,8 @@ import numpy as np
 import pymangle
 import pandas as pd
 from copy import deepcopy
+import logging
+import h5py
 
 from astropy.io import fits
 from astropy.coordinates import search_around_sky
@@ -265,3 +267,14 @@ def gen_randoms(wdir='./data'):
     randoms = randoms[mask]
 
     return randoms.values
+
+
+def save_lightcone(outdir, ra, dec, z, galsnap, galidx, hod_seed=0):
+    outfile = pjoin(outdir, 'obs', f'lightcone{hod_seed}.h5')
+    logging.info(f'Saving lightcone to {outfile}')
+    with h5py.File(outfile, 'w') as f:
+        f.create_dataset('ra', data=ra)            # Right ascension [deg]
+        f.create_dataset('dec', data=dec)          # Declination [deg]
+        f.create_dataset('z', data=z)              # Redshift
+        f.create_dataset('galsnap', data=galsnap)  # Snapshot index
+        f.create_dataset('galidx', data=galidx)    # Galaxy index
